Added a new activity with registration
In the new activity you can enter your first and last name; when you click on the saved button, the data for the current first and last name will be saved and a new UUID will be generated. Each time you sign in, your last entry is saved, so if you don't want to change your GoogleAccount, you shouldn't make any changes.